Move generic functions to the parent module
